deposit instruction now only deposits usdc to vault

+++ programs/yield-aggregator/src/state/vault.rs
@@ -16,15 +16,15 @@ pub struct Vault {
     /// Total shares issued across all users.
     /// Each user's share represents how much of the vault's total underlying they own.
     /// 1 share ~= 1 USDC initially, but as yield accrues, 1 share > 1 USDC.
-    pub total_shares: u128,
+    pub total_shares: u64,
 
     /// Global accumulated yield per share (scaled by SCALER for precision).
     /// Used to calculate user yield since their last update.
-    pub acc_per_share: u128,
+    pub acc_per_share: u64,
 
     /// Total underlying USDC-equivalent value of this vault.
     /// Includes allocations in JupLend, Kamino, and unallocated USDC sitting in the vault.
-    pub total_underlying: u128,
+    pub total_underlying: u64,
 
     /// USDC currently not lent out — held in the vault’s ATA and idle.
     pub unallocated_balance: u64,
@@ -32,18 +32,18 @@ pub struct Vault {
     
     // Allocation Config
     /// Amount of USDC currently deposited in JupLend
-    pub jup_lend_balance: u128,
+    pub jup_lend_balance: u64,
 
     /// Amount of USDC currently deposited in KaminoLend
-    pub kamino_balance: u128,
+    pub kamino_balance: u64,
 
     /// Last recorded value (snapshot) of JupLend allocation.
     /// Used to measure performance (gain/loss) since last update.
-    pub last_jup_value: u128,
+    pub last_jup_value: u64,
 
     /// Last recorded value (snapshot) of KaminoLend allocation.
     /// Used to measure performance (gain/loss) since last update.
-    pub last_kamino_value: u128,
+    pub last_kamino_value: u64,
 
     /// Target allocation percentage 
     pub jup_allocation: u16,    // e.g , 6000 = 60% -> For precision 61.34% = 6134
